refactor: split sandbox-fc control module
